vmware: updateVmwareDc API for updating vmware datacenter details
This adds a new API `updateVmwareDc` that allows admins to update the
VMware datacenter details of a zone. It also recursively updates
the cluster_details and host_details for any username/password updates
as well as updates the `url` detail in cluster_details table and `guid`
detail in the host_details table with any newly provided vcenter
domain/ip. The update API assumes that there is only one vCenter per
zone.

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>