Add more states for attaching from allocated state
for the use case where users want to attach disk in allocated state
but not ready, we need to have more states for keeping track of
attaching state.

Test this by creating a VM with "startvm=false", create a disk and
try attaching it in allocated state. It would give an exception on
latest 4.11 but will be fixed on this patch.