systemvm: new qemu-guest-agent based patching for KVM
This introduces a new patching script for patching systemvms on KVM
using qemu-guest-agent that runs inside the systemvm on startup. This
also removes the vport device which was previously used by the legacy
patching script and instead uses the modern and new uniform guest
agent vport for host-guest communication.

Also updates the sytemvmtemplate build config to use the latest Debian
9.8.0 iso.

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>