server: save SSH keypair exactly as generated by ssh-keygen
On first startup, the management server creates and saves a random
ssh keypair using `ssh-keygen` in the database. The previous code used
to call `trim()` on the string and imposed a restriction of 4k length on
the bytes read from the file.

On latest CentOS7 release (CentOS Linux release 7.6.1810 (Core)) I hit a
bug where router_proxy (ssh) failed due to public key issue and kept
prompting for a passphrase (the default is blank/empty). By adding a
newline to the id_rsa.cloud private key it did not prompt the same and
ssh worked. By performing a visual diff on bytes of the file I found
that the DB version of the keypair and the version that gets put on the
KVM host had a missing `line feed (0xA0)` character. This patch fixes
the issue by ensuring we save the ssh keypair in the database exactly
as it is generated by ssh-keygen. Furthermore, I found that sometimes
the VR startup could fail if the VR took more time to patch, so I
increased the number of retries it previously used to do.

SSH packages details on my CentOS7 env:
libssh2.x86_64                             1.4.3-12.el7_6.2            @updates
openssh.x86_64                             7.4p1-16.el7                @anaconda
openssh-askpass.x86_64                     7.4p1-16.el7                @base
openssh-clients.x86_64                     7.4p1-16.el7                @anaconda
openssh-server.x86_64                      7.4p1-16.el7                @anaconda

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>