router: support multi-homed VMs in VPC
This does not remove VM entries in dbags when hostnames match. The
current codebase already removes entry when a VM is stopped/removed so
we don't need to handle lazy removal. This will allow a VM on
multiple-tiers in a VPC to get dns/dhcp rules as expected.

This also fixes the issue of dhcp_release based on a specific interface.

Fixes #3273

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>