Update dialogue layout and population logic for permissions of template
The older UI was confusing as it showed both project and
accounts. The list of accounts and projects with add/remove
wasn't consistent and instead showed everything.

This patch updates the dialogue layout logic to only allow users
to select accounts or project one at a time. The lists are also
populated such that users see accounts/projects which don't have
permissions to add and accounts/projects that have permissions
to remove.