README: add build dependencies to package.json and document production build properly (#21)
Fixes: #12

With build tooling added to the dev dependencies, it is no longer necessary to install system-wide.
This is particularly useful for production builds, where system-wide package pollution is highly discouraged or even prohibited.

npm@next, vue-cli and ncu may still be useful to developers, but they shouldn't be part of the build instructions. A separate readme update is on the way.

The PR now also includes a readme update to clarify which tools should be installed when.
Production builds do not need global tool installs.

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>