Avoid duplicate alerts when router state changes
When both router of VPC are in MASTER state
then multiple alerts are sent equal to the number of tiers in the VPC.
If the VPC has 3 tiers then 6 alerts will be sent. This is not good
if VPC has more than 10 networks in it.

Instead of checking the router status for all the tiers in the VPC,
just check the status of rotuer for one tier in a VPC so that
multiple deplicate alerts can be avoided