vrouter: reload keepalived instead of restart and fix password server issues when add/remove vpc tier
This fixes few issues below
(1) when add/remove a vpc tier, the keepalived is restarted on both master and backup VRs, which causes some seconds donwtime of vpc.
(2) when remove a vpc tier, the password server is not stopped.
(3) vpc tier is shutdown because no vm is running for 20 minutes, if we start a vm in the vpc tier, it will be implemented. however the guest ip in VPC VRs will be changed , and password server is still running with old guest ip in VRs.