RabbitMQ log enhancement
The handling of the exception can be improved. Example of output: "Failed to create a connection to AMQP server due to null"

The null is caused because it gets the exception message; however, not all exceptions hold a message and therefore we get null instead of the Exception stack, if that log message holds the exception stack trace it would help.