kvm: pre-add 32 PCI controller for hot-plug issue
On newer libvirt/qemu it seems PCI hot-plugging could be an issue as
seen in:

https://www.suse.com/support/kb/doc/?id=000019383
https://bugs.launchpad.net/nova/+bug/1836065

As per the default machine doc, it advises to pre-allocate PCI
controllers on the machine and pcie-to-pci-bridge based controller
for legacy PCI models:
https://libvirt.org/pci-hotplug.html#x86_64-q35

This patch introduces the concept as a workaround until a proper fix is
done (ideally in the upstream libvirt/qemu projects). Until then client
code can add 32 PCI controllers and a pcie-to-pci-bridge controller.

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>
(cherry picked from commit 532b234b19c38553f9ae0535c82857b5ad4f347c)
Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>