Dont add host back after cloudstack-agent service restart
If a host is removed from cloudstack, it will be
added back if we restart the agent service on
the host. It should not be added back if manualy
removed.

Add a global setting "add.host.on.service.restart" with
default value of true. If set to false, the zoneid,
clusterid and podid in agent.properties will be set to
null so that when service is restarted, the agent is not
added back

If true these values will not be set to null so that it
can be still added back on service restart