Check if host is suitable for migraation
If a vm is running on non dedicated host then all the hosts
which are dedicated for other accounts/domain should not be
eligible for migration.

Similarly if a vm is running on dedicated host then all the hosts
which are only dedicated for this account/domain should be
eligible for migration

Steps to reproducde the issue:

1. Dedicate a host to root/admin
2. create a account test, and create a vm
3. Try to migrate the vm
4. Open the dialog to migrate the vm

Expected result:
The dedicated host for root domain should be "Not Suitable" for migration

Actual result:
The ui shows the "dedicated host" as "Suitable"

Test cases:

Create a test account.
Create 1 vm under root domain and 1 under test domain

1. Try to migrate a vm of root domain
2. All hosts should be suitable for migration - PASS
3. Same result for the vm of test domain - PASS

1. Dedicate a different host to root domain
2. Try to migrate vm of root domain
3. All hosts should be suitable for migration since vm is running on
   non dedicated host and the other hosts are either non dedicated
   or dedicated to root domain - PASS
4. Try to migrate vm of test domain. The host which was dedicated above
   should be not suitable for migration as it is dedicated to different
   domain - PASS

1. Try to dedicate the host on which vm of root domain running to root domain
2. Try to migrate the vm of root domain.
3. All hosts should be not suitable since source host is dedicated and no other
   hosts are dedicated -PASS
4. Try to migrate vm of test domain
5. All hosts except the host which is dedicated to root domain should be not
   suitable for migration. - PASS

Do the same above steps but with dedicating a host to test domain instead of root domain