fix NPE in volumes statistics
If volume is migrated between storage pools and "volume.stats.interval"
is set to more frequent interval for exmaple 30 sec.
NullPointerException is thrown, because the hypervisor is looking for
volume on storage pool from which the volume was migrated.
The statistics did not work right - before this fix only one cluster is
lists if we have more clusters from the same type (eg. KVM), and only
one host will return volume stats. Also all volumes are sent to that
host, but they could be on another host. And the logic does not work
correctly.
Also the response was not working correctly for QCOW2 images, because
it's looking for volume's uuid, but in the statistics CS keeps the path
(because of the migrated volumes which UUIDs are changed)