Qemu 2.10 requires `-U` flag to read volume metadata
`Qemu 2.10 added the requirement of a --force-share flag to `qemu-img info` command when reading information about a disk that is in use by a guest`. After this Qemu release, to obtain the attached disk metadata, we need to use a flag to force information reading.
This PR intends to add the required flag to the `qemu-img info` command if the qemu version is 2.10+. We can get the Qemu version in the LibvirtConnection.