HA Agent check regardless of NFS
Currently, KVM HA implementation works only ifthe KVM cluster shares a NFS storage pool. This implementation adds health checks that work without NFS storage. This is done via a Java client that checks Agent status via a webserver.

The additional webserver exposes a simple JSON API which returns a list of Virtual Machines that are running on that host according to libvirt.

This way, KVM HA can verify, via libvirt, VMs status with a HTTP-call to this simple webserver and determine if the host is actually down or if it is just the Java Agent which has crashed.