Allow domain admins to create offering without mentioning domainid
While creating disk/service offering by domains we need to
specify domain id else it wont be created. Provide an enhancement
so that we dont need to pass domainid and it will be taken
from the callers user similar to other api's

Another bug fix

Currently while creating any offering, we need to
specify a domain id or a list of domain id.
If an offering is created on a domain then all its child
domains should see it and all the parents domain
should see it as well but it should not be visible
to sibling domain

isrecursive=true ensures that all child domains can see it
but if a service offering is created for the domain at the leaf
of the domain path then its parent cant see it if isrecurise
is true. so we need to pass isrecursive=false so that the
parent domain can see the child offerings